adds code for including feels like temperature
Adds code to use the feels like temperature in addition to the actual temperature. It is off by default. In ansiweatherrc, it is called via `show_feels_like` true or false, and from the command line by `-h` (short for heat index, since both `f` and `F` are taken and `w` (for wind chill) is taken by wind speed and direction. Addresses issue #133